pyproject.toml
[tool.poetry]
name = "coup"
version = "0.1.0"
description = "A terminal text-based version of the popular game, The Resistance: Coup"
authors = ["Dirk Brand <dirkbrnd@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.5.2"
names = "^0.3.0"
rich = "^13.7.0"
yfinance = "^0.2.40"
langchain = "^0.2.9"
langchain-openai = "^0.1.17"
langchain-core = "^0.2.21"
langgraph = "^0.1.8"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.2"
isort = "^5.8.0"
flake8 = "^3.9.1"
pytest = "^6.2.4"
pytest-asyncio = "^0.15.0"
black = "^22.6.0"


[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
exclude = '''
(
  /(
    | \.git          # root of the project
    | \.venv
  )
)
'''

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
